python/xml2json/requirements.txt
python-magic==0.4.15
xmltodict==0.11.0
